Add more tracing points for data communications

+++ tests/simple/gmac/opencl/gmacThreadVecAdd.cpp
@@ -20,7 +20,7 @@ const size_t blockSize = 32;
 static float **s;
 
 const char *kernel = "\
-__kernel void vecAdd(__global float *c, __global const float *a, __global const float *b, unsigned long size)\
+__kernel void vecAdd(__global float *c, __global const float *a, __global const float *b, unsigned size)\
 {\
     unsigned i = get_global_id(0);\
     if(i >= size) return;\
